ReactTestComponent CustomInput and DescribeRenderer converted to TS
