+++ src/components/SvelteTestComponent/ItRenderer/ItRenderer.jsx
@@ -3,7 +3,6 @@ import cn from 'classnames';
 import styles from '../../ReactTestComponent/ItRenderer/ItRenderer.module.scss';
 import { Draggable } from 'react-beautiful-dnd';
 import SvelteTestStatements from '../../TestCase/SvelteTestStatements';
-import CustomInput from '../CustomInput/CustomInput';
 import {
   addRender,
   addAction,
@@ -69,14 +68,6 @@ const SvelteItRenderer = ({
             onClick={deleteItStatementHandleClick}
             className={cn(styles.itClose, 'far fa-window-close')}
           />
-          {/* <CustomInput
-            key={`input-${id}-${i}`}
-            id={id}
-            label={'The component should...'}
-            placeholder={'Button component renders correctly...'}
-            value={itStatements.byId[id].text}
-            handleChange={handleChangeItStatementText}
-          /> */}
           <div id={styles.itInputContainer}>
             <TextField
               key={`input-${id}-${i}`}
