+++ src/utils/reactTypes.ts
@@ -14,7 +14,7 @@ export interface ReactStatements {
     statements: Statements
   }
   export interface DescribeBlocks {
-    byId: Object;
+    byId: DescribeById;
     allIds: Array<string>;
   }
   export interface ItStatements {
@@ -53,4 +53,11 @@ export interface ReactStatements {
     allIds: Array<string>;
     componentPath: string;
     componentName: string;
+  }
+
+  export interface DescribeById {
+    [key: string]: {
+      id: string,
+      text: string
+    }
   }
